Fixed parsing of long integer numbers (between 10 and 18 digits).

+++ eclipse-project-jdk14/src/com/solab/iso8583/parse/FieldParseInfo.java
@@ -168,12 +168,12 @@ public class FieldParseInfo {
 			//A long covers up to 18 digits
 			if (length < 19) {
 				long l = 0;
-				int power = 1;
+				long power = 1L;
 				for (int i = pos + (length / 2) + (length % 2) - 1; i >= pos; i--) {
 					l += (buf[i] & 0x0f) * power;
-					power *= 10;
+					power *= 10L;
 					l += ((buf[i] & 0xf0) >> 4) * power;
-					power *= 10;
+					power *= 10L;
 				}
 				return new IsoValue(IsoType.NUMERIC, new Long(l), length, null);
 			} else {
